+++ src/jmh/java/com/apptasticsoftware/insynsregistret/InsynsregistretBenchmark.java
@@ -1,4 +1,4 @@
-package com.apptastic.insynsregistret;
+package com.apptasticsoftware.insynsregistret;
 
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.Scope;
