+++ src/test/java/com/apptasticsoftware/insynsregistret/TransactionQueryTest.java
@@ -1,4 +1,4 @@
-package com.apptastic.insynsregistret;
+package com.apptasticsoftware.insynsregistret;
 
 import org.junit.After;
 import org.junit.Before;
